# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  category: string;
  imageUrl: string;
  rating: number;
  reviewCount: number;
  stockQuantity: number;
  inStock: boolean;
  featured: boolean;
  /** @nullable */
  badge?: string | null;
}

export interface Category {
  name: string;
  slug: string;
  productCount: number;
}

export interface StorefrontSummary {
  featuredProducts: Product[];
  totalProducts: number;
  categoryCount: number;
}

export interface Promotion {
  id: number;
  title: string;
  message: string;
  /** @nullable */
  discountText?: string | null;
  /** @nullable */
  code?: string | null;
  ctaLabel: string;
  ctaUrl: string;
}

export interface AdminPromotionInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  message: string;
  /** @nullable */
  discountText?: string | null;
  /** @nullable */
  code?: string | null;
  /** @minLength 1 */
  ctaLabel: string;
  /** @minLength 1 */
  ctaUrl: string;
  active: boolean;
  /** @nullable */
  startsAt?: string | null;
  /** @nullable */
  endsAt?: string | null;
}

export type AdminPromotion = AdminPromotionInput & ({
  id: number;
  /** @nullable */
  startsAt: string | null;
  /** @nullable */
  endsAt: string | null;
  createdAt: string;
});

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  email: string;
  /** @minLength 1 */
  shippingAddress: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export interface Order {
  id: number;
  status: string;
  total: number;
  createdAt: string;
}

export interface AdminProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  imageUrl: string;
  /**
     * @minimum 0
     * @maximum 5
     */
  rating?: number;
  /** @minimum 0 */
  reviewCount?: number;
  /** @minimum 0 */
  stockQuantity: number;
  inStock: boolean;
  featured: boolean;
  /** @nullable */
  badge?: string | null;
}

export type AdminProduct = Product;

export interface AdminOrderItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
}

export interface AdminOrder {
  id: number;
  customerName: string;
  email: string;
  shippingAddress: string;
  status: string;
  total: number;
  createdAt: string;
  items: AdminOrderItem[];
}

export type AdminOrderStatusInputStatus = typeof AdminOrderStatusInputStatus[keyof typeof AdminOrderStatusInputStatus];


export const AdminOrderStatusInputStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface AdminOrderStatusInput {
  status: AdminOrderStatusInputStatus;
}

export interface AdminSummary {
  totalProducts: number;
  inStockProducts: number;
  lowStockProducts: number;
  pendingOrders: number;
  revenue: number;
  recentOrders: AdminOrder[];
}

export type ListProductsParams = {
category?: string;
search?: string;
featured?: boolean;
};