# lib/api-zod/src/generated/types/adminPromotion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminPromotionInput } from './adminPromotionInput';

export type AdminPromotion = AdminPromotionInput & ({
  id: number;
  /** @nullable */
  startsAt: Date | null;
  /** @nullable */
  endsAt: Date | null;
  createdAt: Date;
});