# lib/api-zod/src/generated/types/adminPromotionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminPromotionInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  message: string;
  /** @nullable */
  discountText?: string | null;
  /** @nullable */
  code?: string | null;
  /** @minLength 1 */
  ctaLabel: string;
  /** @minLength 1 */
  ctaUrl: string;
  active: boolean;
  /** @nullable */
  startsAt?: Date | null;
  /** @nullable */
  endsAt?: Date | null;
}