# lib/api-zod/src/generated/types/promotion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Promotion {
  id: number;
  title: string;
  message: string;
  /** @nullable */
  discountText?: string | null;
  /** @nullable */
  code?: string | null;
  ctaLabel: string;
  ctaUrl: string;
}